pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plottheme"
version = "0.1.0"
description = "A robust infrastructure for managing and applying custom matplotlib plot themes."
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "Pillow>=9.0.0",
]

[tool.setuptools.package-data]
plottheme = ["fonts/Swift/*.ttf", "fonts/Swift/*.otf"]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Visualization",
]

[project.urls]
Homepage = "https://github.com/yourusername/plottheme"
